backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
yt-dlp==2024.8.6
python-dotenv>=1.0.0
httpx>=0.27.0
certifi>=2024.0.0
python-multipart>=0.0.9
